Inline a test
